Fix normalized cross-correlation equation, subtract mean.
The formal definition of normalized cross-correlation requires the
subtraction of the mean before summing the squares.  The usual trick
can be used to subtract the mean afterwards, as long as the sums were
computed with sufficient precision.

If the mean is not subtracted, then the metric cannot be used if the
transform scales the image.