Re-enable the cluster merging code.
I don't like cluster merging, it is generally helpful only because
the correct cluster was not found in the first place.  The problem
is that using distance alone is not the best way to identify a cluster,
perhaps the cluster sum should play a role, too.  Or perhaps the number
of slices that the cluster passes through.  Some weighted combination
of these rather than distance alone.