ENH: Use a better parameterization for rotation.
Using versor parameters to describe rotation in 3D registration
is like using sin(theta) instead of just theta for 2D registration.
A delta in a rotation parameter should equate to approximately the
same distance travelled on the surface of a unit sphere, just as a
delta theta for 2D equates to distance travelled around a unit circle.