Force updating of WholeExtent on VTK 5.
There was some difficulty because VTK tends to make the UpdateExtent
sticky, so that it remains the same on subsequent updates if not
explicity set to a new value.  Unfortunately, if the WholeExtent had
shrunk, then the UpdateExtent would be larger than the WholeExtent
and hence would cause a error.