Add the vtkImageConnectivityFilter.
This is an image connectivity filter that works in two modes:
1) with seeds, where a region is grown from each seed.
2) without seeds, where regions are automatically labeled.
By default, images are labelled by the seed scalars (if there are
seeds with scalars), otherwise they are labelled consectitively,
or optionally they can be labelled in order of decreasing region
size.  It is also possible to extract only the largest of the
detected regions.