Add Brent's method to Powell minimizer.
Previously "Powell" only used golden-section for the line search.
Brent's method has the potential to converge much faster for smooth
functions, but a lot of tuning needs to be done before we will see
much improvement!  The problems seem to be 1) proper bracketing isn't
done, so only tiny steps are taken at the beginning, 2) the tolerance
is too large, so only 1 or 2 linear-search steps are taken.  There
also seem to be problems due to lack of smoothness of the objective
function.