Fix a memory overrun in vtkImageRegionIteratorBase.
If the stencil extent was smaller than the image extent, then the code
incremented the span list pointer upon reaching the first voxel within
the stencil extent, even though the pointer was already at the correct
location for the first voxel in the stencil extent.  The extra increment
could lead to a buffer overrun and unpredictable behavior resulting in
random crashes.