Fix empty clipExtent for interpolators
The clip extent provides the intersection of the output extent and the
input extent, taking the kernel size into account.  If it is empty, it
should always be specified the same way: the lower bound should be the
same as the output's lower bound, and the upper bound should be one unit
less than the lower bound (so that the computed size in any dimension is
zero).