Add vtkNelderMeadMinimizer (re-based vtkAmoebaMinimizer)
This is the vtkAmoebaMinimizer from VTK, except that it now has the
vtkFunctionMinimizer as a base class.  One difference from the original
is that when Minimize() is called, it calls GetAmoebaParameterValues()
on each iteration, instead of just once at the end.  This will slightly
reduce the performance, but usually the evaluation of the cost function
will entirely swamp the cost of GetAmoebaParameterValues.