Set metric "cost" to a value between 0 and 1
This makes it easier to compare the metrics and to (eventually) combine
them.  I haven't done Cross Correlation yet because it needs to know the
intensity ranges of the two images, and it does not have that information
yet.  Also, I did not touch the Squared Difference metric, so its cost is
still exactly the squared difference.