Add target image stencil for registration.
The target image mask is resampled to put it into the source image space,
and this is done with nearest-neighbor interpolation.  The resampled mask
is an approximation, and the voxels on the border of the mask will be
within the registration's interpolation kernel neighborhood.  There is
some work to be done in order to fix this issue.