Normalized cross-correlation is undefined for flat field.
The vtkImageNeighborhoodCorrelation metric was misbehaving for images
that had large flat areas.  When the NCC denominator was zero, I was
setting the NCC to 1.0, but the NCC is actually undefined when the
neighborhood is flat.  The correct solution is not to allow these flat
neighborhoods to contribute to the metric, i.e. don't include them in
the count.