Avoid using the reslice->SetInformationInput() method.
This method causes problems when the "information" image has negative
spacing.  Unless the spacing is explicitly set to a negative value via
the reslice->SetOutputSpacing() method, the vtkImageReslice output
spacing will be positive.  For image registration, this means the moving
image has a positive spacing even if the fixed image has negative spacing.
As a result, the voxels aren't stored in the same order, voxel-to-voxel
similarity comparisons no longer work.