Change parameterization for affine transformation.
The previous parameterization used three parameters for the scales and
three parameters for the rotation of these scales (i.e. the directions
of the principal axes).  Unfortunately, when the scales are the same,
the rotation parameters are useless and they cause the optimizer to spin
for a long time while it tries to optimize them.  So, instead I have
parameterized a symmetric matrix, instead of a diagonal matrix with a
rotation.