Fix wrapper linking to Python libraries.
Previously, VTK linked the python libraries transitively so the fact
that VTK_PYTHON_LIBRARIES was empty went unnoticed.  On Jan 16th 2017
VTK stopped linking its wrappers to the python libraries, so it once
again became necessary for us to explicitly link them here, hence we
have to coerce vtkWrapPython.cmake to generate VTK_PYTHON_LIBRARIES
correctly.

Note that it should be possible to completely avoid linking to the
python libraries at compile time and allow the symbols to be resolved
at run time.  That can wait for another day.